Extract cross version fixture for accessing extensions and conventions
